onlineStore
A simple Java program to simulate an online store. 
The program displaying a list of available products and their prices. Then asking the user to choose a product and then enter the product's quantity. The program allowing the user to keep choosing more products and quantities until they enter 0 to indicate they want to end the program. In the end the program telling the user the total amount for the products they have selected.